Finished most modules
- Moved most of the functions to separate modules.
