+++ modules/switchDisplay.js
@@ -1,0 +1,7 @@
+import {bookList, addBook, contact} from '../Script/index.js';
+
+export default function () {
+    bookList.classList.add('not-display');
+    addBook.classList.add('not-display');
+    contact.classList.add('not-display');
+  }
